Move all authentication routes to the auth blueprint
